requirements.txt
asgiref==3.3.1
Django==3.1.6
lxml==4.6.2
Pillow==8.1.0
python-pptx==0.6.18
pytz==2021.1
sqlparse==0.4.1
XlsxWriter==1.3.7
